perf+ux: row-parallel encode, progress bars, --kv and --target flags
Encoding parallelizes across row chunks within each tensor, so one huge
tensor no longer serializes the tail (0.6B IQ encode 25s -> 18.5s, full
core utilization; output verified byte-identical). Measurement and
encode phases show progress bars. --kv {f16,q8_0,q4_0} budgets the KV
cache at its real size and passes cache-type flags to llama-server;
--target 16GB budgets for another Mac by RAM (74% working-set
approximation). Clippy-clean with an explicit allow for the
index-for-index kernel style.

+++ src/quant_iq.rs
@@ -741,8 +741,7 @@ pub fn dec_iq2_s(data: &[u8], out: &mut Vec<f32>) {
                 d * (0.5 + (sc[ib32] >> 4) as f32) * 0.25,
             ];
             for l in 0..4 {
-                let idx = qs[4 * ib32 + l] as usize
-                    | ((((qh[ib32] as usize) << (8 - 2 * l)) & 0x300) as usize);
+                let idx = qs[4 * ib32 + l] as usize | (((qh[ib32] as usize) << (8 - 2 * l)) & 0x300);
                 let grid = IQ2S_GRID[idx].to_le_bytes();
                 let signs = qs[32 + 4 * ib32 + l];
                 for j in 0..8 {
